chore: extract useWithdrawButtonProps hook and fix bugs
